backend/pyproject.toml
[project]
name = "dream-newspaper-backend"
version = "0.1.0"
description = "꿈신문사 - AI 멀티에이전트 꿈 신문 서비스 백엔드"
requires-python = ">=3.13"
dependencies = [
    # 웹 프레임워크
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.9",
    # 데이터베이스
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    # AI
    "anthropic>=0.40.0",
    # 인증
    "python-jose[cryptography]>=3.3.0",
    "passlib>=1.7.4",
    "argon2-cffi>=23.1.0",
    # 스케줄러
    "apscheduler>=3.10.0",
    # 설정 & 유틸
    "pydantic-settings>=2.6.0",
    "pydantic[email]>=2.9.0",
    "httpx>=0.28.0",
    "authlib>=1.3.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
    "supabase>=2.28.0",
    "aiosqlite>=0.22.1",
    "google-genai>=1.0.0",
    "pgvector>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.13"
strict = false
ignore_missing_imports = true
